Increased number of supported zones
Certain parts of the old code (KV pairs handling and few other things)
were setting the limit on the number of zones too low. Currently it
should be able to support up to 28 zones.

Note: core code and logic can handle up to 64 zones, but UI interface is
limiting it right now. Will fix it later.