Created new unified "Station" code based on MasterStation
Previously Master and Remote stations had similar but different code.
Now it is moved to unified "Station" code that could be compiled for
different hardware and for Master or Remote operation.

The code is based on previous "Master" code. New code requires beefier
hardware for Remote, 1284P or 2560 MCU. But it is deemed to be
acceptable because I switched to Moteino Mega as the baseline MCU for
both stations.