Fixed issues with W5500 library
1. W5500 shield running with Moteino Mega MCU is showing some problems
that seem to be related to concurrent activity (interrupts). The change
of disabling SPI interrupts seems to fix it.

2. On soft-reboot of the MCU W5500 sometimes shows errors of being
unable to open socket. It seems to be related to the lack of reset of
the W5500 controller on the main MCU soft-reboot. Adding W5500 reset to
init code appears to fix it.