Work in progress
The code now works correctly on both Remote and Master stations, and it
also allows Master station to act as a Client/Remote - e.g. it is
possible to control and query local channels on Master station via RF.

I think at this stage v1.5 code can be declared Alpha.