Added decent system logging
Added reasonably comprehensive system logging to the local SD card and
remote (to Master), with associated RPRotocol support.

In the process of doing it I converted error messages that previously
went to TRACE to go to the system log.

Actually system log will mirror messages to trace, so it is a super-set
of functionality.