Enabled Schedules view in Logs and backend infra for it.
Enabled Schedules view in Logs and backend infra for it. Also schedules
logs now is the default Logs view.

Schedules Logs view is way faster than Zones or Graph - especially when
there are lots of zones, and it is a lot more meaningful - it shows
which schedules were running and for how long. Zones view is still
available, as detailed drill-down view.