Fixed RF collision condition when executing schedules with remote zones
Fixed RF collision condition when executing schedules with remote zones.

When running a schedule that controls remote (RF-connected) zones, there
was a problem with RF collision. When a zone is turned Off, Remote
station will be sending zone status report to Master to report completed
operation. But at about the same time Master will be trying to turn On
next zone, leading to collision when attempting to send RF packet.
Visible effect of this problem was occasional failure to run some of the
remote zones when executing a schedule.

XBee was gracefully handling such collision, probably due to robust
CSMA-CD and retry algorithms, but unfortunately RFM69 is not as smart,
and RFM69 library (Moteino RF library) does not handle collisions well.

I mitigated this collision problem by making Master add a small (5
seconds) delay between zones in a schedule. This allows enough time for
Remote station to send status reports.

I might need to get back to it later and improve collision handling in
RFM69 library, but for now this mitigation seems to be sufficient to
resolve immediate problem.