Fixed watermeter log generation, also added custom multiplier for watermeter visualization
There was a bug in sdlog.cpp code that prevented watermeter logs to
render in time series - fixed it.

Also added fixed multiplier to UI code to convert water meter units.
System itself records water meter signal in native units of the water
meter (0.1 ccf), but for visualization it is far more convenient to work
with Gallons. Ideally this conversion should be done at backend, but it
requires infrastructure that is not in place.

Doing it in UI is a hack, but will work for now.